% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{add_ribbon_legend}
\alias{add_ribbon_legend}
\title{Add Ribbon Legend Layer to a ggplot}
\usage{
add_ribbon_legend(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  x_major = NULL,
  y_major = NULL,
  x_minor = NULL,
  ymin_minor = NULL,
  ymax_minor = NULL,
  x_scale = identity,
  y_scale = identity,
  global_rescale = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{x_minor, ymin_minor, ymax_minor}{Aesthetics to map plot coordinates
for major and minor glyph components.}
}
\value{
A ggplot2 layer.
}
\description{
This function adds a custom ribbon legend layer to a ggplot object using
the specified aesthetics and parameters.
}
