% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{calculate_min_dist}
\alias{calculate_min_dist}
\title{Calculate the Smallest Distance Across Unique Combinations of Major Axes}
\usage{
calculate_min_dist(data)
}
\arguments{
\item{data}{A data frame containing columns `x_major` and `y_major`, which represent the coordinates
(longitude and latitude) for each point.}
}
\value{
A list containing:
\describe{
  \item{width}{The smallest distance between any two unique points, converted from meters to degrees of latitude.}
  \item{height}{The height corresponding to the width, calculated using the golden ratio (1:1.618).}
}
}
\description{
This function calculates the smallest distance between all unique coordinate combinations of major axes
using the Haversine formula, which measures distances on the surface of a sphere (i.e., Earth).
The function returns the minimum distance in degrees of latitude and a corresponding height that
maintains a glyph ratio of 1:1.618 (the golden ratio).
}
\keyword{internal}
