% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{configure_glyph_data}
\alias{configure_glyph_data}
\title{Setup Glyph Data Based on Geometric Plot Type}
\usage{
configure_glyph_data(data, params, ...)
}
\value{
A modified data frame prepared for plotting glyphs, based on the
geometric plot type.
}
\description{
Setup Glyph Data Based on Geometric Plot Type
}
\keyword{internal}
