% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_segment.R
\docType{data}
\name{GeomGlyphSegment}
\alias{GeomGlyphSegment}
\title{GeomSegmentGlyph-}
\description{
GeomSegmentGlyph-
}
\seealso{
\link[ggplot2]{GeomSegment} from the ggplot2 package.
}
\keyword{internal}
