% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{get_scale}
\alias{get_scale}
\title{Retrieve function from global environment}
\usage{
get_scale(x)
}
\value{
The function retrieved either directly from the list or from the global environment.
}
\description{
This function retrieves a function from a list.
}
\keyword{internal}
