% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{glyph_box}
\alias{glyph_box}
\title{Create reference boxes for glyph plot}
\usage{
glyph_box(data, params)
}
\value{
A data frame with additional columns (`xmin`, `xmax`, `ymin`, `ymax`)
defining the bounding box for each glyph based on the major axes and specified dimensions.
}
\description{
Create reference boxes for glyph plot
}
\keyword{internal}
