% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{glyph_mapping}
\alias{glyph_mapping}
\title{Scaled positional adjustment}
\usage{
glyph_mapping(spatial, scaled_value, length)
}
\value{
numeric vector representing the spatio-temporal transformation of minor axes into
spatial coordinates.
}
\description{
Scaled positional adjustment
}
\keyword{internal}
