% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{glyph_setup_grob}
\alias{glyph_setup_grob}
\title{Convert ggplot2 object into grob}
\usage{
glyph_setup_grob(data, panel_params)
}
\value{
A grob (graphical object) converted from a ggplot2 object.
}
\description{
Convert ggplot2 object into grob
}
\keyword{internal}
