% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{ref_line}
\alias{ref_line}
\title{Calculate reference lines for glyph plot}
\usage{
ref_line(data, params)
}
\value{
A data frame with additional columns (`x`, `y`, `group`) that define
reference lines for each glyph based on the major axes and specified width.
}
\description{
Calculate reference lines for glyph plot
}
\keyword{internal}
