% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{rescale}
\alias{rescale}
\title{Adjust minor axes to to fit within an interval of [-1,1]}
\usage{
rescale(dx)
}
\description{
Adjust minor axes to to fit within an interval of [-1,1]
}
\keyword{internal}
