% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\docType{data}
\name{GeomGlyphRibbon}
\alias{GeomGlyphRibbon}
\alias{GeomGlyphBox}
\alias{GeomGlyphLine}
\alias{GeomGlyphLegend}
\title{GeomGlyphRibbon}
\description{
GeomGlyphRibbon

GeomGlyphBox

GeomGlyphLine

GeomGlyphLegend
}
\keyword{datasets}
