% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\docType{data}
\name{GeomGlyphRibbon}
\alias{GeomGlyphRibbon}
\alias{GeomGlyphBox}
\alias{GeomGlyphLine}
\alias{GeomGlyphLegend}
\title{GeomGlyphRibbon}
\description{
GeomGlyphRibbon

GeomGlyphBox

GeomGlyphLine

GeomGlyphLegend
}
\seealso{
\link[ggplot2]{GeomRibbon} from the ggplot2 package.

\link[cubble]{geom_glyph_box} from the cubble package.

\link[cubble]{geom_glyph_line} from the cubble package.
}
\keyword{internal}
