% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_glyph.R
\name{theme_glyph}
\alias{theme_glyph}
\title{Apply a Custom Glyph Theme to ggplot2 Plots}
\usage{
theme_glyph(font = "sans")
}
\arguments{
\item{font}{A character string specifying the font family to use for the text elements in the plot. The default is "sans".}
}
\value{
A ggplot2 theme object that can be added to ggplot2 plot objects using the `+` operator.
}
\description{
This function customizes the appearance of ggplot2 plots by applying a black-and-white theme with specified font settings for various plot components. It allows for easy customization of the font used throughout the plot while maintaining a clean and professional look.
}
\examples{
library(ggplot2)
# Basic usage with default font
p <- ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_glyph()
print(p)

# Usage with a custom font
p <- ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_glyph(font = "serif")
print(p)

}
