% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{add_glyph_boxes}
\alias{add_glyph_boxes}
\title{Add Glyph Boxes layer to glyph plot}
\usage{
add_glyph_boxes(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  x_major = NULL,
  y_major = NULL,
  x_minor = NULL,
  ymin_minor = NULL,
  ymax_minor = NULL,
  alpha = 1,
  height = ggplot2::rel(2),
  width = ggplot2::rel(2.3),
  color = "grey85",
  linetype = "solid",
  fill = "white",
  inherit.aes = TRUE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer. If \code{NULL}, the default, the data is
inherited from the plot data as specified in the call to \code{ggplot()}.}

\item{x_major, y_major, x_minor, ymin_minor, ymax_minor}{Aesthetics to map plot coordinates
for major and minor glyph components.}

\item{alpha}{The transparency level of the glyph box (ranges between 0 and 1).}

\item{height, width}{The relative height and width of each glyph box.}

\item{color}{The color of the outline of the glyph box.}

\item{linetype}{The type of line to use for the outline of the glyph box.}

\item{fill}{The color used to fill the glyph box.}

\item{...}{Additional parameters.}
}
\value{
A layer object that can be added to a ggplot.
}
\description{
This function introduces a custom layer to a ggplot, employing 'glyph boxes'
to visually represent individual glyph. Users can specify various aesthetics
including alpha, height, width, color, linetype, and fill to customize the appearance.
}
