% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{add_ref_lines}
\alias{add_ref_lines}
\title{Add Reference Lines to glyph plot}
\usage{
add_ref_lines(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  x_major = NULL,
  y_major = NULL,
  x_minor = NULL,
  ymin_minor = NULL,
  ymax_minor = NULL,
  height = ggplot2::rel(2),
  width = ggplot2::rel(2.3),
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer. If \code{NULL}, the default, the data is
inherited from the plot data as specified in the call to \code{ggplot()}.}

\item{x_major, y_major, x_minor, ymin_minor, ymax_minor}{Aesthetics to map plot coordinates
for major and minor glyph components.}
}
\value{
A ggplot2 layer.
}
\description{
This function draw reference lines that include both major and minor division markers.
}
