% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flight Cancellation Data from 2019-2023}
\format{
## `flights`
A data frame with 989355 rows and 45 columns:
\describe{
  \item{FL_DATE}{The date of the flight}
  \item{ORIGIN}{The origin airport for that flight}
  ...
}
}
\usage{
flights
}
\description{
A list of canceled flights that originated from the top 10 airports with the
most canceled flights. The included airports are DEN, MCO, SEA, ATL, DFW,
ORD, LAS, LAX, and PHX.
}
\keyword{datasets}
