% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{glyph_setup_data}
\alias{glyph_setup_data}
\title{Prepare data for geom_glyph_ribbon}
\usage{
glyph_setup_data(data, params, ...)
}
\description{
Prepare data for geom_glyph_ribbon
}
\keyword{internal}
