% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ribbon.R
\name{rescale}
\alias{rescale}
\alias{rescale01y}
\alias{rescale11y}
\title{Rescaling Functions}
\usage{
rescale(dx)

rescale01y(y, yend, ylim = NULL)

rescale11y(y, yend, xlim = NULL)
}
\description{
Adjust minor axes to to fit within an interval of [-1,1]
#' @param x numeric vector
}
\keyword{internal}
