% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{rescale}
\alias{rescale}
\alias{rescale01y}
\alias{rescale11y}
\title{Rescale Functions}
\usage{
rescale(dx)

rescale01y(y, yend, ylim = NULL)

rescale11y(y, yend, xlim = NULL)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A rescaled numeric vector or list of numeric vectors,
with values adjusted to fit within the interval [-1, 1] or [0, 1],
depending on the function used.
}
\description{
Rescale Functions
}
\keyword{internal}
