% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R, R/sugarglider_package.R
\docType{data}
\name{GeomGlyphRibbon}
\alias{GeomGlyphRibbon}
\alias{GeomGlyphBox}
\alias{GeomGlyphLine}
\alias{GeomGlyphLegend}
\alias{sugarglider}
\title{GeomGlyphRibbon}
\description{
This function provides ggplot2 extensions to create glyph maps.
}
\details{
Please see the help pages listed below:

\itemize{
  \item \code{\link{geom_glyph_ribbon}}
  \item \code{\link{geom_glyph_segment}}
}

Also see the vignette for more usage examples:

\code{browseVignettes("sugarglider")}

Please report issues and suggest improvements at Github:

\url{https://maliny12.github.io/sugarglider/}
}
\seealso{
\link[ggplot2]{GeomRibbon} from the ggplot2 package.

\link[cubble]{geom_glyph_box} from the cubble package.

\link[cubble]{geom_glyph_line} from the cubble package.

Useful links:
\itemize{
  \item \url{https://maliny12.github.io/sugarglider/}
  \item \url{https://github.com/maliny12/sugarglider}
  \item Report bugs at \url{https://github.com/maliny12/sugarglider/issues}
}

}
\author{
\strong{Maintainer}: Maliny Po \email{malinypo12@gmail.com} (\href{https://orcid.org/0009-0008-4686-6631}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nathan Yang \email{nathan.s.yang@gmail.com}
}

Other contributors:
\itemize{
  \item H. Sherry Zhang \email{huizezhangsh@gmail.com} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID}) [contributor]
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID}) [contributor]
}

}
\keyword{internal}
