% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_glyph.R
\name{theme_glyph}
\alias{theme_glyph}
\title{Customized Theme for Glyph Plots}
\usage{
theme_glyph()
}
\value{
A ggplot2 theme object with customized settings for glyph plots.
}
\description{
`theme_glyph()` provides a customized theme for glyph maps,
built on top of `theme_map()` from `ggthemes`. It adjusts the plot's appearance, including
the legend position, text styles, and background settings, to create
a clean, visually consistent layout for glyph visualizations.
}
\details{
This theme includes:
\itemize{
  \item Legend positioned inside the plot, at the bottom left corner.
  \item Horizontal legend direction with `monospace` font for text.
  \item Centered plot titles with bold, `monospace` font.
  \item `monospace` fonts for subtitles and captions.
  \item White background for both the panel and plot.
}
}
\examples{
library(ggplot2)
library(ozmaps)

aus_temp |>
  ggplot(aes(x_major = long, y_major = lat,
             x_minor = month, ymin_minor = tmin,
             ymax_minor = tmax)) +
 geom_sf(data = abs_ste, fill = "antiquewhite",
         inherit.aes = FALSE, color = "white") +
 add_glyph_boxes() +
 add_ref_lines() +
 geom_glyph_ribbon() +
 theme_glyph()

}
\seealso{
[ggthemes::theme_map()], [ggplot2::theme()]
}
