% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\docType{data}
\name{train}
\alias{train}
\title{Monthly Train Station Patronage 2023-2024}
\format{
A data frame with each row representing aggregated monthly data per station, containing:
\describe{
  \item{station_name}{Name of the train station.}
  \item{month_year}{Operating month}
  \item{long}{Longitude of the train station.}
  \item{lat}{Latitude of the train station.}
  \item{services}{Number of unique services passing through the station.}
  \item{mode}{Transportation mode, such as 'Metro', 'VLine', or both.}
  \item{min_monthly}{Minimum daily patronage by month.}
  \item{max_monthly}{Maximum daily patronage by month.}
  \item{min_weekday}{Minimum daily patronage on a typical weekday.}
  \item{max_weekday}{Maximum daily patronage on a typical weekday.}
  \item{min_weekend}{Minimum daily patronage on weekends.}
  \item{max_weekend}{Maximum daily patronage on weekends.}
  \item{min_preAM_peak}{Minimum daily patronage during pre-AM peak hours.}
  \item{max_preAM_peak}{Maximum daily patronage during pre-AM peak hours.}
  \item{min_AM_peak}{Minimum daily patronage during AM peak hours.}
  \item{max_AM_peak}{Maximum daily patronage during AM peak hours.}
  \item{min_interpeak}{Minimum daily patronage during interpeak hours.}
  \item{max_interpeak}{Maximum daily patronage during interpeak hours.}
  \item{min_PM_peak}{Minimum daily patronage during PM peak hours.}
  \item{max_PM_peak}{Maximum daily patronage during PM peak hours.}
  \item{min_PM_late}{Minimum daily patronage during late PM hours.}
  \item{max_PM_late}{Maximum daily patronage during late PM hours.}
}
}
\source{
\url{https://discover.data.vic.gov.au/dataset/train-service-passenger-counts}
}
\usage{
data(train)
}
\description{
This dataset provides a comprehensive monthly summary of daily patronage at each
train station in Victoria for the fiscal year 2023-2024. The number of patronage is definded by the
total number of boardings and alightings at each station. This dataset includes detailed breakdowns by day types
such as weekdays, weekends, and different peak periods, enhancing transit management
and planning capabilities. Note the influence of station closures which may skew or omit data
on particular days.
}
\note{
Stations with incomplete entries due to closure are not included in this dataset.
}
\keyword{data}
