% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_glyph_ribbon.R
\name{update_params}
\alias{update_params}
\title{Defined Glyph dimension}
\usage{
update_params(data, params)

update_params(data, params)
}
\value{
A list containing updated `width` and `height` values.

A list containing updated `width` and `height` values.
}
\description{
Adjusts the width and height of glyphs based on the distinct values of
`x_major` and `y_major` in the data.
}
\keyword{internal}
